+++ app/src/androidTest/java/com/agileburo/anytype/features/editor/DeleteBlockTesting.kt
@@ -24,7 +24,7 @@ import com.agileburo.anytype.mocking.MockDataFactory
 import com.agileburo.anytype.presentation.page.PageViewModel
 import com.agileburo.anytype.ui.page.PageFragment
 import com.agileburo.anytype.utils.CoroutinesTestRule
-import com.agileburo.anytype.utils.TestUtils
+import com.agileburo.anytype.utils.TestUtils.withRecyclerView
 import com.bartoszlipinski.disableanimationsrule.DisableAnimationsRule
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.stub
@@ -47,14 +47,14 @@ class DeleteBlockTesting : EditorTestSetup() {
     @get:Rule
     val coroutineTestRule = CoroutinesTestRule()
 
-    val args = bundleOf(PageFragment.ID_KEY to root)
+    private val args = bundleOf(PageFragment.ID_KEY to root)
 
     @Before
     override fun setup() {
         super.setup()
     }
 
-    //region SCENARIO 1
+    //region SCENARIO I
 
     /**
      * SCENARIO I
@@ -231,6 +231,36 @@ class DeleteBlockTesting : EditorTestSetup() {
         )
     }
 
+    @Test
+    fun shouldDeleteSecondH1ByDeletingItsTextThenFocusFirstHeader1WithCursorAtItsEnd() {
+        shouldDeleteSecondBlockByDeletingItsTextThenFocusFirstOneWithCursorAtItsEnd(
+            firstStyle = Block.Content.Text.Style.H1,
+            secondStyle = Block.Content.Text.Style.H1,
+            firstViewId = R.id.headerOne,
+            targetViewId = R.id.headerOne
+        )
+    }
+
+    @Test
+    fun shouldDeleteSecondH2ByDeletingItsTextThenFocusFirstHeader2WithCursorAtItsEnd() {
+        shouldDeleteSecondBlockByDeletingItsTextThenFocusFirstOneWithCursorAtItsEnd(
+            firstStyle = Block.Content.Text.Style.H2,
+            secondStyle = Block.Content.Text.Style.H2,
+            firstViewId = R.id.headerTwo,
+            targetViewId = R.id.headerTwo
+        )
+    }
+
+    @Test
+    fun shouldDeleteSecondH3ByDeletingItsTextThenFocusFirstHeader3WithCursorAtItsEnd() {
+        shouldDeleteSecondBlockByDeletingItsTextThenFocusFirstOneWithCursorAtItsEnd(
+            firstStyle = Block.Content.Text.Style.H3,
+            secondStyle = Block.Content.Text.Style.H3,
+            firstViewId = R.id.headerThree,
+            targetViewId = R.id.headerThree
+        )
+    }
+
     private fun shouldDeleteSecondBlockByDeletingItsTextThenFocusFirstOneWithCursorAtItsEnd(
         firstStyle: Block.Content.Text.Style,
         secondStyle: Block.Content.Text.Style,
@@ -301,7 +331,7 @@ class DeleteBlockTesting : EditorTestSetup() {
         // TESTING
 
         val target = Espresso.onView(
-            TestUtils.withRecyclerView(R.id.recycler).atPositionOnView(2, targetViewId)
+            withRecyclerView(R.id.recycler).atPositionOnView(2, targetViewId)
         )
 
         target.apply {
@@ -317,7 +347,7 @@ class DeleteBlockTesting : EditorTestSetup() {
         verifyBlocking(unlinkBlocks, times(1)) { invoke(params) }
 
         Espresso.onView(
-            TestUtils.withRecyclerView(R.id.recycler).atPositionOnView(1, firstViewId)
+            withRecyclerView(R.id.recycler).atPositionOnView(1, firstViewId)
         ).apply {
             check(ViewAssertions.matches(ViewMatchers.withText("Foo")))
             check(ViewAssertions.matches(ViewMatchers.hasFocus()))
@@ -345,7 +375,250 @@ class DeleteBlockTesting : EditorTestSetup() {
 
     //endregion
 
+    //region SCENARIO II
 
+    /**
+     * SCENARIO II
+     * SHOULD DELETE THE FIRST BLOCK AFTER TITLE, THEN FOCUS THE DOCUMENT'S TITLE
+     */
+
+    @Test
+    fun shouldDeleteFirstParagraphAfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.P,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.textContent,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    @Test
+    fun shouldDeleteFirstH1AfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.H1,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.headerOne,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    @Test
+    fun shouldDeleteFirstH2AfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.H2,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.headerTwo,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    @Test
+    fun shouldDeleteFirstH3AfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.H3,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.headerThree,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    @Test
+    fun shouldDeleteFirstHighlightAfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.QUOTE,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.highlightContent,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    @Test
+    fun shouldDeleteFirstCheckboxAfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.CHECKBOX,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.checkboxContent,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    @Test
+    fun shouldDeleteFirstBulletAfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.BULLET,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.bulletedListContent,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    @Test
+    fun shouldDeleteFirstNumberedAfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.NUMBERED,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.numberedListContent,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    @Test
+    fun shouldDeleteFirstToggleAfterTitleByDeletingItsTextThenFocusTitle() {
+        shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+            firstStyle = Block.Content.Text.Style.TOGGLE,
+            secondStyle = Block.Content.Text.Style.P,
+            firstViewId = R.id.toggleContent,
+            secondViewId = R.id.textContent
+        )
+    }
+
+    private fun shouldDeleteFirstBlockAfterTitleByDeletingItsTextThenFocusTitle(
+        title: String = "Title",
+        firstStyle: Block.Content.Text.Style,
+        secondStyle: Block.Content.Text.Style,
+        firstViewId: Int,
+        secondViewId: Int
+    ) {
+
+        // SETUP
+
+        val a = Block(
+            id = MockDataFactory.randomUuid(),
+            fields = Block.Fields.empty(),
+            children = emptyList(),
+            content = Block.Content.Text(
+                text = "Foo",
+                marks = emptyList(),
+                style = firstStyle
+            )
+        )
+
+        val b = Block(
+            id = MockDataFactory.randomUuid(),
+            fields = Block.Fields.empty(),
+            children = emptyList(),
+            content = Block.Content.Text(
+                text = "Bar",
+                marks = emptyList(),
+                style = secondStyle
+            )
+        )
+
+        val page = Block(
+            id = root,
+            fields = Block.Fields(emptyMap()),
+            content = Block.Content.Smart(
+                type = Block.Content.Smart.Type.PAGE
+            ),
+            children = listOf(a.id, b.id)
+        )
+
+        val document = listOf(page, a, b)
+
+        val events = listOf(
+            Event.Command.UpdateStructure(
+                context = root,
+                id = root,
+                children = listOf(b.id)
+            ),
+            Event.Command.DeleteBlock(
+                context = root,
+                targets = listOf(a.id)
+            )
+        )
+
+        val params = UnlinkBlocks.Params(
+            context = root,
+            targets = listOf(a.id)
+        )
+
+        stubInterceptEvents()
+
+        stubOpenDocument(
+            document = document,
+            details = Block.Details(
+                mapOf(
+                    root to Block.Fields(
+                        mapOf("name" to title)
+                    )
+                )
+            )
+        )
+
+        stubUpdateText()
+
+        stubUnlinkBlocks(params, events)
+
+        val scenario = launchFragment(args)
+
+        // TESTING
+
+        Espresso.onView(
+            withRecyclerView(R.id.recycler).atPositionOnView(0, R.id.title)
+        ).apply {
+            perform(ViewActions.click())
+        }
+
+        // Set cursor programmatically
+
+        scenario.onFragment { fragment ->
+            fragment.recycler.findViewById<TextInputWidget>(R.id.title).setSelection(0)
+        }
+
+        Thread.sleep(100)
+
+
+
+        val target = Espresso.onView(
+            withRecyclerView(R.id.recycler).atPositionOnView(1, firstViewId)
+        )
+
+        target.apply {
+            perform(ViewActions.click())
+        }
+
+        // Delete text from B
+
+        repeat(4) { target.perform(ViewActions.pressKey(KeyEvent.KEYCODE_DEL)) }
+
+        // Check results
+
+        verifyBlocking(unlinkBlocks, times(1)) { invoke(params) }
+
+        Espresso.onView(
+            withRecyclerView(R.id.recycler).atPositionOnView(0, R.id.title)
+        ).apply {
+            check(ViewAssertions.matches(ViewMatchers.withText("Title")))
+            check(ViewAssertions.matches(ViewMatchers.hasFocus()))
+        }
+
+        Espresso.onView(
+            withRecyclerView(R.id.recycler).atPositionOnView(1, secondViewId)
+        ).apply {
+            check(ViewAssertions.matches(ViewMatchers.withText("Bar")))
+        }
+
+        // Check cursor position
+
+        scenario.onFragment { fragment ->
+            val item = fragment.recycler.getChildAt(0)
+            val view = item.findViewById<TextInputWidget>(R.id.title)
+            assertEquals(
+                expected = title.length,
+                actual = view.selectionStart
+            )
+            assertEquals(
+                expected = title.length,
+                actual = view.selectionEnd
+            )
+        }
+
+        // Release pending coroutines
+
+        advance(PageViewModel.TEXT_CHANGES_DEBOUNCE_DURATION)
+    }
+
+    //endregion
 
     // STUBBING & SETUP